requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.3
Pillow==2.5.1
Werkzeug==0.9.6
aniso8601==0.84
arrow==0.4.4
beautifulsoup4==4.3.2
blist==1.3.6
cassandra-driver==2.1.2
feedparser==5.1.3
nose==1.3.4
numpy==1.9.0
pymongo==2.7.2
python-dateutil==2.2
pytz==2014.7
requests==2.4.1
six==1.8.0
